Split-up lib and bin inside crate
This also refactors some modules to be imported as well as exposes a
single entry function `Server::start()` for now.
