+++ Cargo.toml
@@ -18,7 +18,7 @@ categories = ["network-programming", "api-bindings"]
 
 [[bin]] # Bin to run the HelloWorld gRPC server
 name = "criserver"
-path = "src/server.rs"
+path = "src/main.rs"
 
 [dependencies]
 anyhow = "1.0.32"
