+++ src/criapi.rs
@@ -1,0 +1,1 @@
+tonic::include_proto!("criapi");
