+++ src/main.rs
@@ -1,0 +1,7 @@
+use cri::Server;
+
+#[tokio::main]
+async fn main() -> Result<(), Box<dyn std::error::Error>> {
+    Server::start().await?;
+    Ok(())
+}
